requirements.txt
SQLAlchemy==2.0.25
transliterate==1.10.2
python-dotenv==1.0.1